+++ src/main/java/net/chesstango/piazzolla/syzygy/PieceSymmetricWdl.java
@@ -21,7 +21,7 @@ class PieceSymmetricWdl extends TableBase {
         U_INT8_PTR data = new U_INT8_PTR(mappedFile);
         data.incPtr(5);
 
-        int tb_size_white = ei_wtm.init_enc_info(data, 0);
+        long tb_size_white = ei_wtm.init_enc_info(data, 0);
 
         data.incPtr(pieceEntry.num + 1);
 
@@ -60,7 +60,7 @@ class PieceSymmetricWdl extends TableBase {
             i = ei.fill_squares(pos, flip, 0, p, i);
         }
 
-        int idx = ei.encode_piece(p);
+        long idx = ei.encode_piece(p);
 
         byte[] w = ei.precomp.decompress_pairs(idx);
